Rename to AirframeHTTPPlugin
