+++ sbt-airframe/src/sbt-test/sbt-airframe/generate-client/build.sbt
@@ -1,5 +1,5 @@
 libraryDependencies += "org.wvlet.airframe" %% "airframe-http" % "20.2.0"
 
-enablePlugins(AirframePlugin)
+enablePlugins(AirframeHttpPlugin)
 
 airframeHttpPackages ++= Seq("example")
